Users

+++ CommunityApp/Services/UserService.cs
@@ -1,11 +1,13 @@
 ﻿using CommunityApp.Data.Models;
 using CommunityApp.Data.Repositories.Interfaces;
+using Microsoft.AspNetCore.Identity;
 
 namespace CommunityApp.Services
 {
-    public class UserService(IUserRepository repository)
+    public class UserService(IUserRepository repository, UserManager<ApplicationUser> userManager)
     {
         private readonly IUserRepository _repository = repository;
+        private readonly UserManager<ApplicationUser> _userManager = userManager;
 
         public async Task<List<UserDto>> GetAllUsersAsync()
         {
@@ -21,8 +23,24 @@ namespace CommunityApp.Services
 
         public async Task<UserDto?> DeleteUserAsync(string userId)
         {
-            var deletedUser = await _repository.DeleteAsync(userId);
-            return deletedUser;
+            try
+            {
+                var user = await _userManager.FindByIdAsync(userId) ?? throw new ArgumentException("User not found", nameof(userId));
+
+                var claims = await _userManager.GetClaimsAsync(user);
+                if (claims.Any(c => c.Type == "IsAdmin" && c.Value == "true"))
+                {
+                    throw new InvalidOperationException("User is an admin");
+                }
+
+                var deletedUser = await _repository.DeleteAsync(userId);
+                return deletedUser;
+            }
+            catch
+            {
+                // _logger.LogError(ex, "Error assigning manager to community.");
+                throw;
+            }
         }
 
         public async Task<bool> AddIsManagerClaimAsync(string userId)
